Remove IComparable interface.
IComparable was only used in one class, HCSearchResult. Changes in PHP made
it a syntax error to declare an implementation of an interface method that
was more restrictive. This caused confusing errors in features that seemed
to be unrelated to the search engine, so this is being removed for simplicity.